% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateCellNeighborhoodDistanceMatrix.R
\name{calculateCellNeighborhoodDistanceMatrix}
\alias{calculateCellNeighborhoodDistanceMatrix}
\title{Calculate Cell Neighborhood Distance Matrix}
\usage{
calculateCellNeighborhoodDistanceMatrix(
  NeighborhoodCenters,
  cellPCs,
  milo_obj,
  chunk_size = 1000
)
}
\arguments{
\item{NeighborhoodCenters}{A matrix where each row represent PC coordinates of neighborhood centers. Can be generated with \code{calculateNeighborhoodCenterMatrix}.}

\item{cellPCs}{A matrix where each row corresponds to a cell and each column corresponds to a principal component. Can be accessed from \code{seurat_obj@reductions$pca@cell.embeddings[,1:12]} following standard Seurat PCA. Only the first 12 PCs are used in this function.}

\item{milo_obj}{A \code{Milo} class object containing the neighborhood connectivity matrix (\verb{@nhoods}).}

\item{chunk_size}{An integer to specify chunk size in Euclidean distance calculations. Memory scales with this parameter.}
}
\value{
A sparse matrix where each element represents the Euclidean distance between a cell and a neighborhood in which it is included.
}
\description{
This function calculates a sparse matrix representing the Euclidean distance between each cell and the neighborhoods in which it is included.
}
