% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateNeighborhoodCenterMatrix.R
\name{calculateNeighborhoodCenterMatrix}
\alias{calculateNeighborhoodCenterMatrix}
\title{Calculate Neighborhood Center Matrix}
\usage{
calculateNeighborhoodCenterMatrix(milo_obj, cellPCs)
}
\arguments{
\item{milo_obj}{A \code{Milo} class object, containing the neighborhood information, including the neighborhood indices \verb{@nhoodIndex} and the neighborhood connectivity matrix \verb{@nhoods}}

\item{cellPCs}{A matrix where each row corresponds to a cell and each column corresponds to a principal component. The first 12 PCs are used to evaluate neighborhood centers.}
}
\value{
A matrix where each row represents the mean values of the first 12 PCs for all cells in a specific neighborhood. Row names correspond to neighborhood identifiers, and columsn correspond to  mean PC values.
}
\description{
This function calculates a matrix where each row represents a neighborhood. The columns define
principal component coordinates for the neighborhood center.
}
