% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateCellScores.R
\name{col_scoring_ignore0}
\alias{col_scoring_ignore0}
\title{Calculate Weighted Average of Sparse Matrices}
\usage{
col_scoring_ignore0(dgCMat, dgCMat2)
}
\arguments{
\item{dgCMat}{A sparse matrix of phenotype association scores for the neighborhoods.}

\item{dgCMat2}{A sparse matrix of distances to the neighborhoods.}
}
\value{
A numeric vector of weighted averages for each column of the input matrices.
}
\description{
This internal function calculates the weighted average of sparse matrices. The weighting is done
using the softmax function applied to the distances between cells and neighborhood centers.
}
\keyword{internal}
