% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateCellNeighborhoodDistanceMatrix.R
\name{euclideanDistanceChunks}
\alias{euclideanDistanceChunks}
\title{Calculate Euclidean distances between rows of two matrices.}
\usage{
euclideanDistanceChunks(matrix1, matrix2, chunk_size = 1000)
}
\arguments{
\item{matrix1}{A float matrix.}

\item{matrix2}{A float matrix.}

\item{chunk_size}{An integer.}
}
\value{
Matrix of pairwise Euclidean distances between the rows of 'matrix1' and 'matrix2'.
}
\description{
euclidean_distances_chunks calculates the Euclidean distances between rows of two matrices.
Matrices are processed in smaller chunks for memory efficiency.
Note that the number of columns in matrix1 and matrix2 must be the same.
}
