% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterAndSubsampleByCellCount.R
\name{filterAndSubsampleByCellCount}
\alias{filterAndSubsampleByCellCount}
\title{Removes individuals from the Seurat object with too few cells, and subsamples individuals with too many cells.}
\usage{
filterAndSubsampleByCellCount(seurat_obj)
}
\arguments{
\item{seurat_obj}{A Seurat object with a meta.data field called 'projid', corresponding to the individual identifier}
}
\value{
a filtered Seurat object.
}
\description{
Individuals are excluded if their number of cells (cellcount) is in the bottom quartile. Conversely, individuals with
more than median + (2*IQR) cells are subsampled. For subsampling, the number of cells to retain is drawn
from the distribution of cell counts from the remaining individuals.
}
