% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removePoorlyMixedCells.R
\name{removePoorlyMixedCells}
\alias{removePoorlyMixedCells}
\title{Remove cells with poor mixture score from Seurat object}
\usage{
removePoorlyMixedCells(seurat_obj, cellPCs, col_name, threshold)
}
\arguments{
\item{seurat_obj}{A Seurat object for which you want poorly mixed cells removed.}

\item{cellPCs}{A matrix of cells (rows) and PCA embeddings (columns). Can be extracted with seurat_obj@reductions$pca@cell.embeddings\link{,1:12}}

\item{col_name}{Define the column name to evaluate mixture. Can be individual, or condition, or a covariate.}

\item{threshold}{A positive real number, defining the minimum LISI required for cell retention.}
}
\value{
A filtered Seurat object
}
\description{
This function calculates the Local Inverse Simpson's Index (LISI)
as a metric of cell mixture. Specifically, LISI measures the effective sample
size in the local embedding space. Poorly mixed space is defined by the
threshold parameter (recommended value at 3).
}
\details{
Note that this function contains a commented debug code block for visualizing
the distribution of LISI scores across cells.
}
