% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runNeightborhoodAssociationTest.R
\name{runNeighborhoodAssociationTest}
\alias{runNeighborhoodAssociationTest}
\title{Test neighborhoods for associations with metadata test variables.}
\usage{
runNeighborhoodAssociationTest(milo_obj, varsToTest, varsToCorrect, col_name)
}
\arguments{
\item{milo_obj}{A Milo object for which makeNhoods and countCells has already been run.}

\item{varsToTest}{A list of variable to test. Ex. c("amyloid","tangles")}

\item{varsToCorrect}{A list of covariates to account for. Ex. c("fixation_interval","msex","pmi","study")}

\item{col_name}{The name of the colnames in the nhoodCounts(milo_obj). Ex "projid".}
}
\value{
A list of data frames, one per test variable. Each data frame will be a table, one row per neighborhood, of enrichment results.
}
\description{
Test neighborhoods for associations with metadata test variables.
}
