% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateCellScores.R
\name{softmax}
\alias{softmax}
\title{Softmax function}
\usage{
softmax(x, temperature = 1)
}
\arguments{
\item{x}{A numeric vector of values.}

\item{temperature}{A positive number controlling the sharpness of the output probabilities.}
}
\value{
A numeric vector the same length as the input vector.
}
\description{
This function computes the softmax of a numeric vector.
}
\keyword{internal}
